UPSTREAM: iio: distance: add devantech us ranger srf04
This patch adds support for the ultrasonic ranger srf04 of devantech.

This device is measuring the distance of objects in a range between 1 cm
and 3 meters and a theoretical resolution of 3 mm.

There are two GPIOs used:
  - trigger: set as output to the device when the measurement should start
  - echo: set by the device when the ultrasonic wave is sent out and reset
    when the echo is recognized; this needs to be an interrupt input

The time between setting and resetting the echo pin is the time the
waveform needed for one round trip. This time is recorded in the interrupt
handler.

The distance is calculated in the read function by using the ultrasonic
speed at 20 degrees celsius which is about 343 m/s.

 Conflicts:
	drivers/iio/proximity/Makefile

Change-Id: I570e4008e1db87c7af9a9107260fda48bfb6542a
Signed-off-by: Andreas Klinger <ak@it-klinger.de>
Signed-off-by: Jonathan Cameron <jic23@kernel.org>
(cherry picked from commit feda2840040fcf87d826b0aec945a98222e656ba)
Signed-off-by: Ziyuan Xu <xzy.xu@rock-chips.com>

rockchip f65a3417e6937730e052262860a70e80844b9383